adfpay_website/requirement.txt
asgiref==3.3.4
dj-database-url==0.5.0
Django==3.1.3
django-ckeditor==6.0.0
django-cors-headers==3.7.0
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
django-social-share==2.2.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.7.1
gunicorn==20.0.4
phonenumbers==8.12.24
Pillow==8.2.0
psycopg2==2.8.6
PyJWT==2.1.0
pytz==2020.4
sqlparse==0.4.1
uuid==1.30
whitenoise==5.2.0
